refactor: :recycle: отделить helpers и ядро преобразования формы
Структурные договоры helpers больше не зависят от выведенных типов. BaseForm и публичный преобразователь используют единое независимое ядро, сохраняя прежний API.
