+++ packages/core/metadata/forms/elements/contextMenu/helper.ts
@@ -1,10 +1,12 @@
-import { ContextMenu } from "./types"
+type ContentWithChildren = {
+  readonly childItems?: readonly unknown[]
+}
 
 export const getContextMenuName = (parentElement: { name: string }): string => {
   return `${parentElement.name}КонтекстноеМеню`
 }
 
-export const isHasContent = (data: ContextMenu | undefined): boolean => {
+export const isHasContent = (data: ContentWithChildren | undefined): boolean => {
   if (!data) return false
 
   if ((data.childItems?.length ?? 0) > 0) return true
