+++ packages/core/metadata/forms/elements/extendedTooltip/helper.ts
@@ -1,12 +1,10 @@
-import { ExtendedTooltip } from "./types"
-
 export const getExtendedTooltipName = (parentElement: { name: string }): string => {
   return `${parentElement.name}РасширеннаяПодсказка`
 }
 
 const EXCLUDED_FIELDS = ["name", "itemType"]
 
-export const isHasContent = (data: ExtendedTooltip): boolean => {
+export const isHasContent = (data: object): boolean => {
   const keys = Object.keys(data)
   const hasOtherFields = keys.some((key) => !EXCLUDED_FIELDS.includes(key))
   return hasOtherFields
