+++ packages/core/metadata/forms/elements/viewStatusAddition/helper.ts
@@ -1,9 +1,7 @@
-import { SingleViewStatusAddition } from "./types"
-
 export const getViewStatusAdditionName = (parentElement: { name: string }): string => {
   return `${parentElement.name}СостояниеПросмотра`
 }
 
-export const isHasContent = (data: SingleViewStatusAddition): boolean => {
+export const isHasContent = (data: object): boolean => {
   return Object.keys(data).length > 0
 }
