+++ tools/dependency-cruiser/test/local-cycle-boundaries.test.mjs
@@ -35,3 +35,26 @@ test("парные реализации зависят от leaf contracts", () 
   assert.doesNotMatch(read("packages/core/xml/import/saxesParser.ts"), /from "\.\/importer"/u)
   assert.doesNotMatch(read("packages/platform/src/sessions/nodeRuntime.ts"), /from "\.\/manager"/u)
 })
+
+test("helpers элементов не импортируют выведенные types.ts", () => {
+  for (const element of [
+    "autoCommandBar",
+    "contextMenu",
+    "searchControlAddition",
+    "viewStatusAddition",
+    "extendedTooltip",
+  ]) {
+    assert.doesNotMatch(read(`packages/core/metadata/forms/elements/${element}/helper.ts`), /from "\.\/types"/u)
+  }
+})
+
+test("baseForm использует независимое ядро", () => {
+  assert.doesNotMatch(
+    read("packages/core/metadata/forms/clientApplicationForm/baseForm.ts"),
+    /from "\.\/fromYAMLToXML"/u
+  )
+  assert.doesNotMatch(
+    read("packages/core/metadata/forms/clientApplicationForm/convertYAMLToXML.ts"),
+    /from "\.\/baseForm"/u
+  )
+})
